Simplify without_echo() C API
No need to export this in C, I think.
