+++ src/init.c
@@ -2,6 +2,7 @@
 #include <Rinternals.h>
 
 #include "keypress.h"
+#include "keypress-internal.h"
 
 static const R_CallMethodDef callMethods[]  = {
   {"keypress", (DL_FUNC) &keypress, 1},
@@ -18,8 +19,4 @@ void R_init_keypress(DllInfo *dll) {
 
   R_RegisterCCallable("keypress", "keypress", (DL_FUNC) keypress);
   R_RegisterCCallable("keypress_read", "keypress_read", (DL_FUNC) keypress_read);
-  R_RegisterCCallable("save_term_status", "save_term_status", (DL_FUNC) save_term_status);
-  R_RegisterCCallable("restore_term_status", "restore_term_status", (DL_FUNC) restore_term_status);
-  R_RegisterCCallable("set_term_echo", "set_term_echo", (DL_FUNC) set_term_echo);
-
-  }
+}
